+++ src/github/gh-api.ts
@@ -12,11 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {context, getOctokit} from '@actions/github';
-import {PullsCreateResponseData} from '@octokit/types';
 import * as core from '@actions/core';
 
 import * as store from '../store';
+
+import {components} from '@octokit/openapi-types';
+import {RequestError} from '@octokit/request-error';
+import {context, getOctokit} from '@actions/github';
+
+type PullsCreateResponseData = components['schemas']['pull-request'];
 
 export interface IGitHubApi {
   repoDefaultBranch: () => Promise<string>;
@@ -60,7 +64,7 @@ export class GitHubApi implements IGitHubApi {
   }
 
   async repoDefaultBranch(): Promise<string> {
-    const {data: repo} = await this.octokit.repos.get({
+    const {data: repo} = await this.octokit.rest.repos.get({
       owner: context.repo.owner,
       repo: context.repo.repo
     });
@@ -79,7 +83,7 @@ export class GitHubApi implements IGitHubApi {
     title: string;
     body: string;
   }): Promise<PullsCreateResponseData> {
-    const {data: pullRequest} = await this.octokit.pulls.create({
+    const {data: pullRequest} = await this.octokit.rest.pulls.create({
       owner: context.repo.owner,
       repo: context.repo.repo,
       head: branchName,
@@ -123,15 +127,20 @@ export class GitHubApi implements IGitHubApi {
     reviewer: string
   ): Promise<boolean> {
     try {
-      const result = await this.octokit.pulls.requestReviewers({
+      const result = await this.octokit.rest.pulls.requestReviewers({
         owner: context.repo.owner,
         repo: context.repo.repo,
         pull_number: pullRequestNumber,
         reviewers: [reviewer]
       });
 
+      if (!result.data.requested_reviewers) {
+        core.warning(`requested_reviewers is empty`);
+        return false;
+      }
+
       const requested_reviewers = result.data.requested_reviewers.map(
-        user => user.login
+        user => user?.login
       );
 
       if (!requested_reviewers.includes(reviewer)) {
@@ -139,9 +148,10 @@ export class GitHubApi implements IGitHubApi {
         return false;
       }
     } catch (error) {
-      core.warning(
-        `Unable to set PR reviewer ${reviewer}, got error: ${error.message}`
-      );
+      core.warning(`Unable to set PR reviewer ${reviewer}`);
+      if (error instanceof Error) {
+        core.warning(`error: ${error.message}`);
+      }
       return false;
     }
 
@@ -183,12 +193,17 @@ export class GitHubApi implements IGitHubApi {
     team: string
   ): Promise<boolean> {
     try {
-      const result = await this.octokit.pulls.requestReviewers({
+      const result = await this.octokit.rest.pulls.requestReviewers({
         owner: context.repo.owner,
         repo: context.repo.repo,
         pull_number: pullRequestNumber,
         team_reviewers: [team]
       });
+
+      if (!result.data.requested_teams) {
+        core.warning(`requested_teams is empty`);
+        return false;
+      }
 
       const requested_teams = result.data.requested_teams.map(t => t.slug);
 
@@ -197,9 +212,10 @@ export class GitHubApi implements IGitHubApi {
         return false;
       }
     } catch (error) {
-      core.warning(
-        `Unable to set PR team reviewer ${team}, got error: ${error.message}`
-      );
+      core.warning(`Unable to set PR team reviewer ${team}`);
+      if (error instanceof Error) {
+        core.warning(`Got error: ${error.message}`);
+      }
       return false;
     }
 
@@ -215,22 +231,23 @@ export class GitHubApi implements IGitHubApi {
     core.info(`Adding labels: ${labels.join(',')}`);
 
     try {
-      await this.octokit.issues.addLabels({
+      await this.octokit.rest.issues.addLabels({
         owner: context.repo.owner,
         repo: context.repo.repo,
         issue_number: pullRequestNumber,
         labels
       });
     } catch (error) {
-      core.warning(
-        `Unable to add all labels to PR, got error: ${error.message}`
-      );
+      core.warning(`Unable to add all labels to PR`);
+      if (error instanceof Error) {
+        core.warning(`error: ${error.message}`);
+      }
     }
   }
 
   async createLabelIfMissing(labelName: string): Promise<boolean> {
     try {
-      const label = await this.octokit.issues.getLabel({
+      const label = await this.octokit.rest.issues.getLabel({
         owner: context.repo.owner,
         repo: context.repo.repo,
         name: labelName
@@ -240,10 +257,12 @@ export class GitHubApi implements IGitHubApi {
 
       return true;
     } catch (error) {
-      if (error.status === 404) {
-        core.debug('Label not found');
+      if (error instanceof RequestError) {
+        if (error.status === 404) {
+          core.debug('Label not found');
 
-        return await this.createLabel(labelName);
+          return await this.createLabel(labelName);
+        }
       }
 
       return false;
@@ -252,7 +271,7 @@ export class GitHubApi implements IGitHubApi {
 
   async createLabel(labelName: string): Promise<boolean> {
     try {
-      const label = await this.octokit.issues.createLabel({
+      const label = await this.octokit.rest.issues.createLabel({
         owner: context.repo.owner,
         repo: context.repo.repo,
         name: labelName,
@@ -264,9 +283,10 @@ export class GitHubApi implements IGitHubApi {
 
       return true;
     } catch (error) {
-      core.warning(
-        `Unable to create label "${labelName}", got error: ${error.message}`
-      );
+      core.warning(`Unable to create label "${labelName}"`);
+      if (error instanceof Error) {
+        core.warning(`error: ${error.message}`);
+      }
       return false;
     }
   }
@@ -276,7 +296,7 @@ export class GitHubApi implements IGitHubApi {
     body: string
   ): Promise<boolean> {
     try {
-      const result = await this.octokit.issues.createComment({
+      const result = await this.octokit.rest.issues.createComment({
         owner: context.repo.owner,
         repo: context.repo.repo,
         issue_number: pullRequestNumber,
@@ -289,9 +309,10 @@ export class GitHubApi implements IGitHubApi {
 
       return true;
     } catch (error) {
-      core.warning(
-        `Unable to create comment for PR ${pullRequestNumber}, got error: ${error.message}`
-      );
+      core.warning(`Unable to create comment for PR ${pullRequestNumber}`);
+      if (error instanceof Error) {
+        core.warning(`error: ${error.message}`);
+      }
       return false;
     }
   }