+++ src/tasks/main.ts
@@ -15,6 +15,10 @@
 import * as core from '@actions/core';
 import * as glob from '@actions/glob';
 
+import * as git from '../git/git-cmds';
+import * as gitAuth from '../git/git-auth';
+import * as store from '../store';
+
 import {commit} from '../git/git-commit';
 import {createWrapperInfo} from '../wrapperInfo';
 import {createWrapperUpdater} from '../wrapperUpdater';
@@ -22,9 +26,6 @@ import {GitHubOps} from '../github/gh-ops';
 import {IGitHubApi} from '../github/gh-api';
 import {Inputs} from '../inputs';
 import {Releases} from '../releases';
-import * as git from '../git/git-cmds';
-import * as gitAuth from '../git/git-auth';
-import * as store from '../store';
 
 export class MainAction {
   private inputs: Inputs;
@@ -137,12 +138,22 @@ export class MainAction {
         core.endGroup();
 
         core.startGroup('Checking whether any file has been updated');
-        const modifiedFiles = await git.gitDiffNameOnly();
+        let modifiedFiles = await git.gitDiffNameOnly();
         core.debug(`Modified files count: ${modifiedFiles.length}`);
         core.debug(`Modified files list: ${modifiedFiles}`);
         core.endGroup();
 
         if (modifiedFiles.length) {
+          // Running the `wrapper` task a second time ensures that the wrapper jar itself
+          // and the wrapper scripts get updated if a new version is available (happens infrequently).
+          // https://docs.gradle.org/current/userguide/gradle_wrapper.html#sec:upgrading_wrapper
+          core.startGroup('Updating Wrapper (2nd update)');
+          await updater.update();
+          modifiedFiles = await git.gitDiffNameOnly();
+          core.debug(`Modified files count: ${modifiedFiles.length}`);
+          core.debug(`Modified files list: ${modifiedFiles}`);
+          core.endGroup();
+
           core.startGroup('Verifying Wrapper');
           await updater.verify();
           core.endGroup();
@@ -196,7 +207,7 @@ export class MainAction {
     } catch (error) {
       // setFailed is fatal (terminates action), core.error
       // creates a failure annotation instead
-      core.setFailed(`❌ ${error.message}`);
+      core.setFailed(`❌ ${error instanceof Error && error.message}`);
     }
   }