+++ src/tasks/post.ts
@@ -14,9 +14,10 @@
 
 import * as core from '@actions/core';
 
-import {IGitHubApi} from '../github/gh-api';
 import * as gitAuth from '../git/git-auth';
 import * as store from '../store';
+
+import {IGitHubApi} from '../github/gh-api';
 
 export class PostAction {
   private githubApi: IGitHubApi;
@@ -32,7 +33,10 @@ export class PostAction {
       await gitAuth.cleanup();
       await this.reportErroredReviewers();
     } catch (error) {
-      core.debug(`post action task failed with: ${error.message}`);
+      core.debug('Post action task failed');
+      if (error instanceof Error) {
+        core.debug(`error: ${error.message}`);
+      }
     }
   }